+++ Riot/Modules/RoomKeyRequest/RoomKeyRequestViewController.m
@@ -16,7 +16,6 @@
 
 #import "RoomKeyRequestViewController.h"
 
-#import "AppDelegate.h"
 #import "Riot-Swift.h"
 
 @interface RoomKeyRequestViewController () <KeyVerificationCoordinatorBridgePresenterDelegate>