+++ Riot/Modules/SetPinCode/EnterPinCode/EnterPinCodeViewState.swift
@@ -23,9 +23,10 @@ enum EnterPinCodeViewState {
     case choosePin              //  creating pin for the first time, enter for first
     case confirmPin             //  creating pin for the first time, confirm
     case pinsDontMatch          //  pins don't match
-    case unlock            //  after pin has been set, enter pin to unlock
+    case unlock                 //  after pin has been set, enter pin to unlock
     case wrongPin               //  after pin has been set, pin entered wrongly
     case wrongPinTooManyTimes   //  after pin has been set, pin entered wrongly too many times
     case forgotPin              //  after pin has been set, user tapped forgot pin
     case confirmPinToDisable    //  after pin has been set, confirm pin to disable pin
+    case inactive               //  inactive state, only used when app is not active
 }